have finished triangle saver to DB

+++ Lab5/Point.cs
@@ -5,11 +5,11 @@ namespace Lab5
     [Serializable]
     public class Point
     {
-        public double X { get; set; }
-        public double Y { get; set; }
+        public int X { get; set; }
+        public int Y { get; set; }
 
         public Point(){}
-        public Point(double x, double y)
+        public Point(int x, int y)
         {
             X = x;
             Y = y;
